Added aux and householder modules, began GPU implementation

+++ toeplitz.cu
@@ -1,7 +1,7 @@
 #include "toeplitz.h"
 #include <stdio.h>
 
-int BuildToeplitz(double *t_matrix, int dim, int diag_cnt)
+int BuildToeplitz(float *t_matrix, int dim, int diag_cnt)
 {
     // Determine parameter validity
     if(diag_cnt >= 2*dim || diag_cnt == 0 || diag_cnt % 2 == 0)
@@ -10,20 +10,19 @@ int BuildToeplitz(double *t_matrix, int dim, int diag_cnt)
         return -1;
     }
     
-    double *diagConsts = (double *)malloc(diag_cnt*sizeof(*diagConsts));
+    float *diagConsts = (float *)malloc(diag_cnt*sizeof(*diagConsts));
 
     // Assigns random numbers to the diagonal constants
-    for(size_t i = 0; i < diag_cnt; i++)
+    for(size_t i = 0; i < (size_t)diag_cnt; i++)
     {
-        diagConsts[i] = ((double)rand() / (double)RAND_MAX);
-        //printf("diagConsts[%u]=%f\n", i, diagConsts[i]);
+        diagConsts[i] = ((float)rand() / (float)RAND_MAX);
     }
 
     int diagRadius = diag_cnt - (diag_cnt / 2) - 1;
 
-    for(size_t y = 0; y < dim; y++)
+    for(size_t y = 0; y < (size_t)dim; y++)
     {
-        for(size_t x = 0; x < dim; x++)
+        for(size_t x = 0; x < (size_t)dim; x++)
         {
             const size_t idx = x + dim*y;
             
@@ -42,13 +41,13 @@ int BuildToeplitz(double *t_matrix, int dim, int diag_cnt)
     return 0;
 }
 
-void PrintMatrix(double *matrix, int m, int n)
+void PrintMatrix(float *matrix, int m, int n)
 {
-    for(unsigned long i = 0; i < m; i++)
+    for(unsigned long i = 0; i < (size_t)m; i++)
     {
         printf("|");
 
-        for(unsigned long j = 0; j < n; j++)
+        for(unsigned long j = 0; j < (size_t)n; j++)
         {
             const size_t idx = j + m*i;
             printf(" %1.2f", matrix[idx]);
