+++ toeplitz.h
@@ -2,9 +2,9 @@
 #define TOEPLITZ_H
 
 /* Builds a Toeplitz matrix */
-int BuildToeplitz(double *t_matrix, int dim, int diag_cnt);
+int BuildToeplitz(float *t_matrix, int dim, int diag_cnt);
 
 /* Outputs matrix */
-void PrintMatrix(double *matrix, int m, int n);
+void PrintMatrix(float *matrix, int m, int n);
 
 #endif
